refactor: move change email components and styles to its own directory
