+++ apps/src/accounts/AccountInformation/index.tsx
@@ -11,7 +11,7 @@ import {Heading2} from '@cdo/apps/componentLibrary/typography';
 import {getAuthenticityToken} from '@cdo/apps/util/AuthenticityTokenStore';
 import i18n from '@cdo/locale';
 
-import ChangeEmailModal from '../ChangeEmailModal';
+import ChangeEmailModal from '../ChangeEmail/ChangeEmailModal';
 
 import {AccountInformationProps} from './types';
